requirements.txt
pycrypto
requests
passlib
beautifulsoup4
